Runtime reducer estimator fixes

+++ scalding-core/src/main/scala/com/twitter/scalding/reducer_estimation/package.scala
@@ -1,4 +1,4 @@
 package object reducer_estimation {
-  def median(xs: Seq[Long]): Option[Long] = xs.sorted.lift(xs.length / 2)
-  def mean(xs: Seq[Long]): Option[Long] = if (xs.isEmpty) None else Some(xs.sum / xs.length)
+  def median(xs: Seq[Double]): Option[Double] = xs.sorted.lift(xs.length / 2)
+  def mean(xs: Seq[Double]): Option[Double] = if (xs.isEmpty) None else Some(xs.sum / xs.length)
 }
